feat: revision 0.0.20 (#161)
* feat: add frontend gly-engine-micro

* feat: raise error when failed compile game

* fix: http not working LG 2024 #155

* fix: http not working LG 2024 #155 (better comment)

* docs: about some ginga issues #155

* feat: add tmm carousel to gitngore

* feat: new class in ginga http

* fix: new http class ginga

* fix: started focused ginga

* docs: ginga_bug_capabilities

* feat: improves in http #151

* feat: re-add --screen in core ginga

* feat: add core html5_ginga and moved core native html

* docs: Create SECURITY.md

* fix: swapad params std.draw.image in html5

* fix: param orders in native_image_draw

* feat: start sample stream

* feat: working on support media.video

* fix: pipeline in video

* feat: improve argparse in cli

* add flag --enginecdn in cli build

* feat: no background on core html5_ginga

* docs: fixed 2games link and not use lite version #160

* docs: usage prefix lua executavel

* feat: add support to lg webos6 #80

* docs: improve keys in asteroids sample

* ci: add websites for stream sample

* fix: build in ginga-ncl using atobify

* ci: fixes in luau-analyze

* docs: move security.md to .github

* feat: allow mock.video

* feat: add ts types in npmjs #153

* feat: add math.wave #152

* fix: tests cases

* docs: templates lua and js

* docs: improve color list

* feat: update links in README.md

* feat: bump core-native-html5 0.1.2

* fix: drop luaunit and improve tests ginga http

* fix: core ginga pass in luau-analyze linter

* feat: bump core-native-html5 0.1.3

* fix: bundler no variable require

* fix: luau analyze in love

* feat: bump 0.0.20

* docs: add tests for math.wave

* docs: add @micro core to doxygen

* feat: bump doxygen ver 0.0.20

* docs: add CONTRIBUTING guide

+++ src/cli/commands/build.lua
@@ -1,6 +1,8 @@
 local zeebo_buildsystem = require('src/cli/tools/buildsystem')
 local zeebo_fs = require('src/lib/cli/fs')
 local util_fs = require('src/lib/util/fs')
+local env_build = require('src/env/build')
+local atobify = require('src/cli/tools/atobify')
 
 local function build(args)
     args.dist = util_fs.path(args.dist).get_fullfilepath()
@@ -13,20 +15,28 @@ local function build(args)
         args.core = 'html5'
     end
 
+    if args.core == 'html5_ginga' then
+        args.fengari = true
+    end
+
     zeebo_fs.clear(args.dist)
     zeebo_fs.mkdir(args.dist..'_bundler/')
+
+    local atob = env_build.need_atobify(args)
 
     local build_game = zeebo_buildsystem.from({core='game', bundler=true, dist=args.dist})
         :add_core('game', {src=args.game, as='game.lua', prefix='game_', assets=true})
 
     local build_core = zeebo_buildsystem.from(args)
+        :add_rule('please use flag -'..'-enterprise to use commercial modules', 'core=html5_ginga', 'enterprise=false')
         :add_rule('please use flag -'..'-enterprise to use commercial modules', 'core=ginga', 'enterprise=false')
-        :add_rule('please use flag -'..'-enterprise to use commercial modules', 'fengari=true', 'enterprise=false')
         :add_rule('please use flag -'..'-gpl3 to use free software modules', 'gamepadzilla=true', 'gpl3=false')
         --
         :add_core('none')
         --
         :add_core('lite', {src='src/engine/core/lite/main.lua'})
+        --
+        :add_core('micro', {src='src/engine/core/micro/main.lua'})
         --
         :add_core('native', {src='src/engine/core/native/main.lua'})
         --
@@ -37,33 +47,26 @@ local function build(args)
         :add_step('love '..args.dist, {when=args.run})
         --
         :add_core('ginga', {src='ee/engine/core/ginga/main.lua'})
-        :add_file('ee/engine/meta/ginga/main.ncl')
-        :add_step('ginga dist/main.ncl', {when=args.run})
+        :add_meta('ee/engine/meta/ginga/ncl.mustache', {as='main.ncl'})
+        :add_step('ginga dist/main.ncl '..env_build.screen_ginga(args), {when=args.run})
         --
         :add_core('html5', {src='src/engine/core/native/main.lua', force_bundler=true})
-        :add_file('src/engine/core/html5/core-native-html5.js')
-        :add_file('src/engine/core/html5/core-media-html5.js')
-        :add_file('src/engine/core/html5/driver-wasmoon.js', {when=not args.fengari})
-        :add_file('ee/engine/core/html5/driver-fengari.js', {when=args.fengari})
-        :add_file('third_party/json/rxi.lua', {as='jsonrxi.lua', when=args.fengari})
         :add_file('assets/icon80x80.png')
         :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
         --
         :add_core('html5_lite', {src='src/engine/core/lite/main.lua', force_bundler=true})
-        :add_file('src/engine/core/html5/core-native-html5.js')
-        :add_file('src/engine/core/html5/core-media-html5.js')
-        :add_file('src/engine/core/html5/driver-wasmoon.js', {when=not args.fengari})
-        :add_file('ee/engine/core/html5/driver-fengari.js', {when=args.fengari})
-        :add_file('third_party/json/rxi.lua', {as='jsonrxi.lua', when=args.fengari})
+        :add_file('assets/icon80x80.png')
+        :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
+        --
+        :add_core('html5_micro', {src='src/engine/core/micro/main.lua', force_bundler=true})
+        :add_file('assets/icon80x80.png')
+        :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
+        --
+        :add_core('html5_ginga', {src='src/engine/core/native/main.lua', force_bundler=true})
         :add_file('assets/icon80x80.png')
         :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
         --
         :add_core('html5_tizen', {src='src/engine/core/native/main.lua', force_bundler=true})
-        :add_file('src/engine/core/html5/core-native-html5.js')
-        :add_file('src/engine/core/html5/core-media-html5.js')
-        :add_file('src/engine/core/html5/driver-wasmoon.js', {when=not args.fengari})
-        :add_file('ee/engine/core/html5/driver-fengari.js', {when=args.fengari})
-        :add_file('third_party/json/rxi.lua', {as='jsonrxi.lua', when=args.fengari})
         :add_file('assets/icon80x80.png')
         :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
         :add_meta('src/engine/meta/html5_tizen/config.xml')
@@ -71,18 +74,23 @@ local function build(args)
         :add_step('cd '..args.dist..';~/tizen-studio/tools/ide/bin/tizen.sh package -t wgt;true')
         --
         :add_core('html5_webos', {src='src/engine/core/native/main.lua', force_bundler=true})
-        :add_file('src/engine/core/html5/core-native-html5.js')
-        :add_file('src/engine/core/html5/core-media-html5.js')
-        :add_file('src/engine/core/html5/driver-wasmoon.js', {when=not args.fengari})
-        :add_file('ee/engine/core/html5/driver-fengari.js', {when=args.fengari})
-        :add_file('third_party/json/rxi.lua', {as='jsonrxi.lua', when=args.fengari})
         :add_file('assets/icon80x80.png')
         :add_meta('src/engine/core/html5/index.mustache', {as='index.html'})
         :add_meta('src/engine/meta/html5_webos/appinfo.json')
         :add_step('webos24 $(pwd)/dist', {when=args.run})
+        --
+        :add_common_func(atobify.builder('engine_code', args.dist..'main.lua', args.dist..'index.js'), {when=atob and not args.enginecdn})
+        :add_common_func(atobify.builder('game_code', args.dist..'game.lua', args.dist..'index.js'), {when=atob})
+        :add_common_func(zeebo_fs.lazy_del(args.dist..'main.lua'), {when=atob or args.enginecdn})
+        :add_common_func(zeebo_fs.lazy_del(args.dist..'game.lua'), {when=atob})
 
-    build_game:run()
-    local ok, message =  build_core:run()
+    local ok, message = build_game:run()
+
+    if not ok then
+        return false, message
+    end
+
+    ok, message =  build_core:run()
     
     zeebo_fs.rmdir(args.dist..'_bundler/')
